Refactorizacion
Refactorización del código fuente original acomodando el orden correcto de ejecución del mismo además de organizar en carpetas acorde a su uso de casa script, de igual formar se actualizaron acciones o funciones obsoletas, se ha documentado el código en español de forma general
