+++ Prueba_Tecnica_Desarrollador_Unity/Assets/Scripts/Camera/CameraPinchToZoom.cs.meta
