+++ Prueba_Tecnica_Desarrollador_Unity/Assets/Scripts/Enviroment.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1be9da8042d229d47b52c0b59e079b7f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
