+++ Prueba_Tecnica_Desarrollador_Unity/Assets/Scripts/Enviroment/Brick.cs
@@ -1,8 +1,9 @@
 ﻿using UnityEngine;
-using System.Collections;
 
 public class Brick : MonoBehaviour
 {
+    public float Health = 70f;
+    //cuando colisiona con un objeto realiza el calculo de daño
     void OnCollisionEnter2D(Collision2D col)
     {
         if (col.gameObject.GetComponent<Rigidbody2D>() == null) return;
@@ -11,8 +12,9 @@ public class Brick : MonoBehaviour
         if (damage >= 10)
             GetComponent<AudioSource>().Play();
         Health -= damage;
+        //si el daño es igual o menor a cero destruye el gameobject
         if (Health <= 0) Destroy(this.gameObject);
     }
 
-    public float Health = 70f;
+    
 }
