+++ Prueba_Tecnica_Desarrollador_Unity/Assets/Scripts/Enviroment/ParallaxScrolling.cs.meta
