+++ Prueba_Tecnica_Desarrollador_Unity/Assets/Scripts/Player/Bird.cs.meta
