+++ Prueba_Tecnica_Desarrollador_Unity/Assets/Scripts/Player/Constants.cs
