+++ Prueba_Tecnica_Desarrollador_Unity/Assets/Scripts/Player/Destroyer.cs
@@ -2,7 +2,7 @@
 using System.Collections;
 
 public class Destroyer : MonoBehaviour {
-    
+    //trigger si tiene contacto ave o puerco o estructura, desrtuir game object
     void OnTriggerEnter2D(Collider2D col)
     {
         string tag = col.gameObject.tag;
