+++ Prueba_Tecnica_Desarrollador_Unity/Assets/Scripts/Player/Enums.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Assets.Scripts
+﻿namespace Assets.Scripts
 {
     public enum SlingshotState
     {
