+++ Prueba_Tecnica_Desarrollador_Unity/Assets/Scripts/Player/Enums.cs.meta
