+++ Prueba_Tecnica_Desarrollador_Unity/Assets/Scripts/Player/GameManager.cs.meta
